Converts news feed from redux to react-query
