+++ package.json
@@ -14,6 +14,7 @@
     "node-sass": "^4.13.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-query": "^3.15.2",
     "react-redux": "^7.1.3",
     "react-scripts": "4.0.3",
     "redux": "^4.0.5",
@@ -29,7 +30,8 @@
     "eject": "react-scripts eject",
     "preversion": "npm test",
     "version": "git add .",
-    "postversion": "git push && git push --tags"
+    "postversion": "git push && git push --tags",
+    "lint": "eslint . --ext .ts"
   },
   "eslintConfig": {
     "extends": "react-app"
