+++ src/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+import { QueryClientProvider, QueryClient } from 'react-query';
 
 import { createReduxStore } from './utils/configureStore';
 import { iconLibrarySetup } from './utils/iconConfig';
@@ -10,9 +11,13 @@ import './styles/main.sass';
 const store = createReduxStore();
 iconLibrarySetup();
 
+const queryClient = new QueryClient()
+
 ReactDOM.render(
     <Provider store={store}>
-        <App />
+        <QueryClientProvider client={queryClient}>
+            <App />
+        </QueryClientProvider>
     </Provider>,
     document.getElementById('root'),
 );
